+++ R/predict.aldvmm.R
@@ -1,11 +1,9 @@
 #' Predict Method for Adjusted Limited Dependent Variable Mixture Model Fits
 #'
-#' The generic function
+#' The method \code{predict.aldvmm} for the generic function
 #' \ifelse{html}{\code{\link[stats]{predict}}}{\code{stats::predict()}} calls
-#' \ifelse{html}{\code{\link[aldvmm]{aldvmm.pred}}}{\code{aldvmm.pred()}} to
-#' predict outcomes and their standard errors in new data using
-#' \ifelse{html}{\code{\link[aldvmm]{aldvmm.pred}}}{\code{aldvmm.pred()}} and
-#' \ifelse{html}{\code{\link[aldvmm]{aldvmm.sefit}}}{\code{aldvmm.sefit()}}.
+#' \ifelse{html}{\code{\link[aldvmm]{aldvmm.pred}}}{\code{aldvmm.pred()}} and \ifelse{html}{\code{\link[aldvmm]{aldvmm.sefit}}}{\code{aldvmm.sefit()}} to
+#' predict outcomes and their standard errors in new data.
 #'
 #' @param object an object inheriting from class 'aldvmm'.
 #' @param newdata a data frame, list or environment (or object coercible to a