+++ aldvmm.Rcheck/00_pkg_src/aldvmm/R/aldvmm.mm.R
@@ -1,10 +1,13 @@
 #' Creating Design Matrices
 #'
 #' \ifelse{html}{\code{\link[aldvmm]{aldvmm.mm}}}{\code{aldvmm.mm()}} creates
-#' two design matrices, one of the model of component distributions
+#' a list of two design matrices, one of the model of component distributions
 #' (\code{"beta"}) and one of the model of probabilities of component
 #' membership (\code{"delta"}).
-#'
+#' 
+#' @param mf a data frame created by \ifelse{html}{\code{\link[stats]{model.frame}}}{\code{stats::model.frame}} including the variables used in formula supplied to \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{\code{aldvmm()}} plus additional attributes, including "terms" for an object of class "terms" derived from formula.
+#' @param Formula an object of class \code{"Formula"} created by \ifelse{html}{\code{\link[Formula]{Formula}}}{\code{Formula::Formula}} based on the \code{formula} supplied to \ifelse{html}{\code{\link[aldvmm]{aldvmm}}}{
+#'   \code{aldvmm()}}.
 #' @inheritParams aldvmm
 #' @inheritParams aldvmm.ll
 #'
@@ -13,10 +16,9 @@
 #'   \ifelse{html}{\code{\link[stats]{model.matrix}}}{\code{stats::model.matrix()}}
 #'    to create design matrices for models of component distributions
 #'   (\code{"beta"}) and probabilities of component membership (\code{"delta"})
-#'   based on \code{'formula'} supplied to
-#'   \ifelse{html}{\code{\link[aldvmm]{aldvmm.ll}}}{\code{aldvmm.ll()}}. The
+#'   based on a \code{'Formula'} object created by \ifelse{html}{\code{\link[Formula]{Formula}}}{\code{Formula::Formula}} and a model frame created by \ifelse{html}{\code{\link[stats]{model.frame}}}{\code{stats::model.frame}}. The
 #'   design matrix for probabilities of group membership is only created if
-#'   more than one components are specified  in \code{'ncmp'}.
+#'   more than one components are specified in \code{'ncmp'}.
 #'
 #' @return a named list of numeric matrices. \item{\code{beta}}{a numeric
 #'   design matrix for the model of component distributions.}
@@ -25,45 +27,22 @@
 #'
 #' @export
 
-aldvmm.mm <- function(data,
-                      formula,
+aldvmm.mm <- function(mf,
+                      Formula,
                       ncmp,
                       lcoef) {
   
-  # Remove rows with missing values
-  #--------------------------------
-  
-  complete <- stats::complete.cases(data[, all.vars(formula)])
-  if (FALSE %in% complete) {
-    data <- data[complete, ]
-  }
-  
-  # List of formulae for distributions and multinomial logit
-  #---------------------------------------------------------
-  
-  # If no pipe (|) delimiter is used in the formula, the same formula will be
-  # used for distributions and the multinomial logit parts.
-  
-  if (grepl("\\|", as.character(formula)[3])) {
-    formvec <- paste0(as.character(formula)[2], 
-                      " ~ ", 
-                      unlist(strsplit(as.character(formula)[3], 
-                                      split = "\\|")))
-    names(formvec) <- lcoef
-    formlist <- lapply(formvec, function(x) stats::as.formula(x))
-  } else {
-    formlist <- lapply(lcoef, function(x) formula)
-    names(formlist) <- lcoef
-  }
-  
-  # List of Model matrices
-  #-----------------------
-  
-  # Make list of model matrices for beta and delta
   if (ncmp > 1) {
-    mm <- lapply(formlist, function(x) stats::model.matrix(x, data))
+    if (length(Formula)[2] > 1) {
+      mm <- list(stats::model.matrix(Formula, data = mf, rhs = 1),
+                 stats::model.matrix(Formula, data = mf, rhs = 2))
+    } else {
+      mm <- list(stats::model.matrix(Formula, data = mf, rhs = 1),
+                 stats::model.matrix(Formula, data = mf, rhs = 1))
+    }
+    names(mm) <- lcoef
   } else {
-    mm <- list(stats::model.matrix(formlist[[1]], data))
+    mm <- list(stats::model.matrix(Formula, data = mf, rhs = 1))
     names(mm) <- lcoef[1]
   }